test: move 07-tendermint client state testing to use light client module entry points (#6074)
